alpacahack/daily/encoding-basics/pyproject.toml
[project]
name = "encoding-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pycryptodome>=3.23.0",
]
